+++ CHANGELOG.md
@@ -70,6 +70,7 @@
 - Added `tower` implementation for `HttpClient`
 - Added `helix::Response::new` and `helix::Response::with_data` to make it possible to create your own responses.
 - Added `is_branded_content` and `content_classification_labels` to `Get Channel Information` and `Modify Channel information`
+- Added `channel.update` v2 EventSub event
 
 ### Fixed